% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links_pagina.R
\name{eliminar_vazios}
\alias{eliminar_vazios}
\title{Title}
\usage{
eliminar_vazios(x)
}
\arguments{
\item{x}{vetor de caracteres}
}
\value{
o mesmo vetor sem os elementos vazios ("")
}
\description{
Title
}
