% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_noticia.R
\name{ler_noticia}
\alias{ler_noticia}
\title{Tira informações da noticia}
\usage{
ler_noticia(sessao = html_session("http://www.valor.com.br"), url)
}
\arguments{
\item{sessao}{Uma sessao}

\item{url}{O link para uma noticia}
}
\value{
Uma lista com o html, o titulo, os autores, as tags e o texto da noticia
}
\description{
Tira informações da noticia
}
\examples{
url_noticia <- 'http://www.valor.com.br/brasil/5179248/ipc-s-tem-alta-de-033-em-outubro-aponta-fgv'
ler_noticia(url = url_noticia)

}
