% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links_pagina.R
\name{links_pagina}
\alias{links_pagina}
\title{Pega noticias da sessao}
\usage{
links_pagina(sessao)
}
\arguments{
\item{sessao}{Uma sessao criada por \code{\link{html_session}}}
}
\value{
Uma string com os links da pagina no css
}
\description{
Pega noticias da sessao
}
\examples{
html_session("http://www.valor.com.br") \%>\% links_pagina()
}
