% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Realizar Login no site do Valor Economico}
\usage{
login(sessao, email = NULL, senha = NULL)
}
\arguments{
\item{sessao}{uma sessao criada por \link{\code{webdriver::Session}}}

\item{email}{e-mail da conta realizar login no site do valor}

\item{senha}{senha para realizar login no site do valor}
}
\value{
Uma sessao conectada ao usuario
}
\description{
Realizar Login no site do Valor Economico
}
\examples{
\dontrun{
pjs <- webdriver::run_phantomjs()
sessao <- webdriver::Session$new(port = pjs$port)
login(sessao, 'fulano@provedor.org', 'senha')
}
}
