% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantomJS.R
\name{phantomJS}
\alias{phantomJS}
\title{Pegar caminho e versao do phantomJS}
\usage{
phantomJS(version = "2.1.1", check = TRUE, verbose = FALSE)
}
\arguments{
\item{version}{...}
}
\value{
caminho do phantomJS
}
\description{
Pegar caminho e versao do phantomJS
}
