% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_noticia.R
\name{print.noticia}
\alias{print.noticia}
\title{Método para imprimir noticia}
\usage{
\method{print}{noticia}(x, ...)
}
\arguments{
\item{x}{Objeto que vai imprimir}

\item{...}{Outros argumento repassados para print}
}
\value{
O mesmo objeto passado em \code{x}
}
\description{
Método para imprimir noticia
}
