% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procurar_valor.R
\name{procurar_valor}
\alias{procurar_valor}
\title{Procurar noticias no site do jornal Valor Economico}
\usage{
procurar_valor(sessao, termo, paginas = 5)
}
\arguments{
\item{sessao}{Sessao}

\item{termo}{Termo buscado}

\item{paginas}{numero de paginas que serao retornadas}
}
\value{
Um vetor com os links das noticias encontradas
}
\description{
Procurar noticias no site do jornal Valor Economico
}
\examples{
\dontrun{
procurar_valor('desemprego')
}

}
