% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Realizar Login no site do Valor}
\usage{
login(email, senha)
}
\arguments{
\item{email}{e-mail da conta realizar login no site do valor}

\item{senha}{senha para realizar login no site do valor}
}
\value{
Uma sessao conectada ao usuario
}
\description{
Realizar Login no site do Valor
}
\examples{

}
