% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantomJS.R
\name{phantomJS}
\alias{phantomJS}
\title{Pegar caminho e versao do phantomJS}
\usage{
phantomJS(version = "2.1.1", check = TRUE, verbose = FALSE)
}
\arguments{
\item{version}{versao do phantomJS que sera buscada}

\item{check}{Novas versoes de phantomJS devem ser buscadas e baixadas?}

\item{verbose}{A funcao deve ser prolixa?}
}
\value{
caminho do phantomJS
}
\description{
Pegar caminho e versao do phantomJS
}
