% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procurar_valor.R
\name{procurar_valor}
\alias{procurar_valor}
\title{Procurar noticias no site do jornal Valor Economico}
\usage{
procurar_valor(termo, sessao = rvest::html_session("http://www.valor.com.br"),
  paginas = 5)
}
\arguments{
\item{termo}{Termo buscado}

\item{sessao}{Sessao}

\item{paginas}{numero de paginas que serao retornadas}
}
\value{
Uma sessao com resultados da busca
}
\description{
Procurar noticias no site do jornal Valor Economico
}
\examples{
\dontrun{
procurar_valor('desemprego')
}

}
